Add test for basic query
