+++ api/githubLogin.js
@@ -15,6 +15,27 @@ const store = new KnexSessionStore({
 });
 
 export function setUpGitHubLogin(app) {
+  if (!GITHUB_CLIENT_ID) {
+    console.warn('GitHub client ID not passed; login won\'t work.'); // eslint-ignore-line no-console
+    return;
+  }
+
+  const gitHubStrategyOptions = {
+    clientID: GITHUB_CLIENT_ID,
+    clientSecret: GITHUB_CLIENT_SECRET,
+    callbackURL: process.env.NODE_ENV !== 'production' ?
+      'http://localhost:3000/login/github/callback' :
+      'http://www.githunt.com/login/github/callback',
+  };
+
+  passport.use(new GitHubStrategy(gitHubStrategyOptions,
+    (accessToken, refreshToken, profile, cb) => {
+      cb(null, profile);
+    }));
+
+  passport.serializeUser((user, cb) => cb(null, user));
+  passport.deserializeUser((obj, cb) => cb(null, obj));
+
   app.use(session({
     secret: 'your secret',
     resave: true,
@@ -37,16 +58,3 @@ export function setUpGitHubLogin(app) {
     res.redirect('/');
   });
 }
-
-const gitHubStrategyOptions = {
-  clientID: GITHUB_CLIENT_ID,
-  clientSecret: GITHUB_CLIENT_SECRET,
-  callbackURL: process.env.NODE_ENV !== 'production' ? 'http://localhost:3000/login/github/callback' : 'http://www.githunt.com/login/github/callback',
-};
-
-passport.use(new GitHubStrategy(gitHubStrategyOptions, (accessToken, refreshToken, profile, cb) => {
-  cb(null, profile);
-}));
-
-passport.serializeUser((user, cb) => cb(null, user));
-passport.deserializeUser((obj, cb) => cb(null, obj));
